# Faster implementation for SentencePieceExtractor