# What does this PR do?

This PR is to improve `SentencePieceExtractor` extract method performance, which took several minutes when targeting vocabularies of tens of thousands of words.

For 44,876 words, it used to take 290 seconds, but now it takes 0.2 seconds.
I've added a simple test, but let me know if you need anything else.
The experimental conditions and code are as follows.

result
```
vocabulary length: 44876, max word length: 16
normal: 290.2607123851776 secs
improved: 0.18401217460632324 secs
```

code
```python
import pickle
import time
from collections import defaultdict
from typing import List

vocab = pickle.load(open('vocab.pkl', 'rb'))

def normal(vocab: dict) -> List[str]:
    merges = []
    for piece_l in vocab.keys():
        for piece_r in vocab.keys():
            merge = f"{piece_l}{piece_r}"
            piece_id = vocab.get(merge, None)
            if piece_id:
                merges += [(piece_l, piece_r, piece_id)]
    return merges

def improved(vocab: dict) -> List[str]:
    merges = []
    prefixes = dict()
    for word in vocab.keys():
        for i in range(len(word)):
            prefixes[word[: i + 1]] = {word} | prefixes.setdefault(word[: i + 1], set())

    for word in vocab.keys():
        if len(prefixes[word]) > 1:
            for candidate in prefixes[word]:
                if word != candidate:
                    if candidate[len(word) :] in vocab:
                        piece_id = vocab.get(candidate, None)
                        merges += [(word, candidate[len(word) :], piece_id)]
    return merges

print(f'vocabulary length: {len(vocab)}, max word length: {max(len(word) for word in vocab.keys())}')
start = time.time()
result_normal = normal(vocab)
print(f'normal: {time.time() - start} secs')

start = time.time()
result_improved = improved(vocab)
print(f'improved: {time.time() - start} secs')

# confirm that results match
assert sorted(result_normal, key=lambda val: val[2]) == sorted(result_improved, key=lambda val: val[2])
```


## Before submitting
- [ ] This PR fixes a typo or improves the docs (you can dismiss the other checks if that's the case).
- [x] Did you read the [contributor guideline](https://github.com/huggingface/transformers/blob/main/CONTRIBUTING.md#start-contributing-pull-requests),
      Pull Request section?
- [ ] Was this discussed/approved via a Github issue or the [forum](https://discuss.huggingface.co/)? Please add a link
      to it if that's the case.
- [ ] Did you make sure to update the documentation with your changes? Here are the
      [documentation guidelines](https://github.com/huggingface/transformers/tree/main/docs), and
      [here are tips on formatting docstrings](https://github.com/huggingface/transformers/tree/main/docs#writing-source-documentation).
- [x] Did you write any new necessary tests?


## Who can review?

@n1t0, @LysandreJik