@Narsil 2022-05-18
Hi @e-mon ,

Thanks for looking into this. 
Is this code used often ? This code was written hastily and was supposed to only run once (since we can save the `tokenizer.json` within the `tokenizers` library which should load again pretty fast.

Since we're looking at optimizing this code I propose another version which seems even faster:
https://gist.github.com/Narsil/a6b927c4973d4d0a63b1765cfff38e55

(I am using a smaller vocab for faster testing, since the slow is really excruciatingly slow).