@lancercat 2023-04-27
Or shall we just parallize it?

```
# coding=utf-8
# Copyright 2018 The HuggingFace Inc. team.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
"""
Utilities to convert slow tokenizers in their fast tokenizers counterparts.

All the conversions are grouped here to gather SentencePiece dependencies outside of the fast tokenizers files and
allow to make our dependency on SentencePiece optional.
"""

import warnings
from typing import Dict, List, Tuple

from tokenizers import AddedToken, Regex, Tokenizer, decoders, normalizers, pre_tokenizers, processors
from tokenizers.models import BPE, Unigram, WordPiece
import multiprocessing
from functools import partial
from .utils import requires_backends
def merge_core(vocab,vocab_scores,piece_ls):
    merges = []
    for piece_l in piece_ls:
        for piece_r in vocab.keys():
            merge = f"{piece_l}{piece_r}"
            piece_score = vocab_scores.get(merge, None)
            if piece_score:
                merges += [(piece_l, piece_r, piece_score)]
    return merges;


class SentencePieceExtractor:
    """
    Extractor implementation for SentencePiece trained models. https://github.com/google/sentencepiece
    """

    def __init__(self, model: str):
        requires_backends(self, "sentencepiece")
        from sentencepiece import SentencePieceProcessor

        self.sp = SentencePieceProcessor()
        self.sp.Load(model)

    def extract(self, vocab_scores=None) -> Tuple[Dict[str, int], List[Tuple]]:
        """
        By default will return vocab and merges with respect to their order, by sending `vocab_scores` we're going to
        order the merges with respect to the piece scores instead.
        """
        sp = self.sp
        vocab = {sp.id_to_piece(index): index for index in range(sp.GetPieceSize())}
        if vocab_scores is not None:
            vocab_scores, reverse = dict(vocab_scores), True
        else:
            vocab_scores, reverse = vocab, False
        pool_obj = multiprocessing.Pool();
        merges=pool_obj.map(partial(merge_core,vocab,vocab_scores),vocab.keys())
        # Merges
        # merges = []
        # for piece_l in vocab.keys():
        #     for piece_r in vocab.keys():
        #         merge = f"{piece_l}{piece_r}"
        #         piece_score = vocab_scores.get(merge, None)
        #         if piece_score:
        #             merges += [(piece_l, piece_r, piece_score)]
        merges = sorted(merges, key=lambda val: val[2], reverse=reverse)
        merges = [(val[0], val[1]) for val in merges]
        return vocab, merges

```